ConnectionSVG utils and test
without tests
